make compile and configure scripts safer

+++ configure.sh
@@ -1,3 +1,9 @@
 #!/bin/bash
-rm -rf build
+
+set -euo pipefail
+
+if [ -d build ]; then
+	rm -rf build
+fi
+
 cmake -S . -B build -G "Unix Makefiles" -DCMAKE_BUILD_TYPE=Release
